# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AttachmentOutcome": {
        "additionalProperties": false,
        "description": "Position only: names and content are never echoed.",
        "properties": {
          "kept": {
            "title": "Kept",
            "type": "boolean"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "reason": {
            "anyOf": [
              {
                "enum": [
                  "unsupported_type",
                  "spoofed_type",
                  "active_content",
                  "encrypted_pdf",
                  "too_large",
                  "too_many_pixels",
                  "too_many_pages",
                  "malformed",
                  "malware_detected",
                  "scan_failed",
                  "storage_failed",
                  "timeout"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "position",
          "kept",
          "reason"
        ],
        "title": "AttachmentOutcome",
        "type": "object"
      },
...
      "ContactOut": {
        "properties": {
          "channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "channel",
          "value"
        ],
        "title": "ContactOut",
        "type": "object"
      },
...
      "EvidenceOut": {
        "description": "Metadata only: no object key, no URL. Content is fetched through the download broker.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "evidence_id": {
            "format": "uuid",
            "title": "Evidence Id",
            "type": "string"
          },
          "mime_type": {
            "title": "Mime Type",
            "type": "string"
          },
          "sanitation_state": {
            "title": "Sanitation State",
            "type": "string"
          },
          "scan_state": {
            "title": "Scan State",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "evidence_id",
          "display_name",
          "mime_type",
          "size_bytes",
          "sanitation_state",
          "scan_state",
          "created_at"
        ],
        "title": "EvidenceOut",
        "type": "object"
      },
...
      "FieldErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "field": {
            "title": "Field",
            "type": "string"
          }
        },
        "required": [
          "field",
          "code"
        ],
        "title": "FieldErrorBody",
        "type": "object"
      },
...
      "ProblemDetails": {
        "description": "OpenAPI description of the error body (the wire shape is built by ``problem_response``).",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldErrorBody"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "detail"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
      "QueueItemOut": {
        "properties": {
          "concern_category": {
            "title": "Concern Category",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "evidence_count": {
            "title": "Evidence Count",
            "type": "integer"
          },
          "has_contact": {
            "title": "Has Contact",
            "type": "boolean"
          },
          "open_follow_ups": {
            "title": "Open Follow Ups",
            "type": "integer"
          },
          "project_slug": {
            "title": "Project Slug",
            "type": "string"
          },
          "report_id": {
            "format": "uuid",
            "title": "Report Id",
            "type": "string"
          },
          "risk_level": {
            "title": "Risk Level",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "status_updated_at": {
            "format": "date-time",
            "title": "Status Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "report_id",
          "project_slug",
          "concern_category",
          "risk_level",
          "status",
          "version",
          "created_at",
          "status_updated_at",
          "has_contact",
          "evidence_count",
          "open_follow_ups"
        ],
        "title": "QueueItemOut",
        "type": "object"
      },
      "QueuePageOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/QueueItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "QueuePageOut",
        "type": "object"
      },
...
      "ReportDetailOut": {
        "properties": {
          "concern_category": {
            "title": "Concern Category",
            "type": "string"
          },
          "contact": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContactOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/StatusEventOut"
            },
            "title": "Events",
            "type": "array"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceOut"
            },
            "title": "Evidence",
            "type": "array"
          },
          "follow_ups": {
            "items": {
              "$ref": "#/components/schemas/FollowUpOut"
            },
            "title": "Follow Ups",
            "type": "array"
          },
          "has_contact": {
            "title": "Has Contact",
            "type": "boolean"
          },
          "project_slug": {
            "title": "Project Slug",
            "type": "string"
          },
          "report_id": {
            "format": "uuid",
            "title": "Report Id",
            "type": "string"
          },
          "risk_level": {
            "title": "Risk Level",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "status_updated_at": {
            "format": "date-time",
            "title": "Status Updated At",
            "type": "string"
          },
          "track_record": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TrackRecordOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "report_id",
          "project_slug",
          "concern_category",
          "risk_level",
          "status",
          "version",
          "created_at",
          "status_updated_at",
          "has_contact",
          "description",
          "events",
          "follow_ups",
          "evidence",
          "track_record",
          "contact"
        ],
        "title": "ReportDetailOut",
        "type": "object"
      },
      "ReportReceipt": {
        "additionalProperties": false,
        "description": "Shown once. The client must display the code and never put it in a URL or storage.",
        "properties": {
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/AttachmentOutcome"
            },
            "title": "Attachments",
            "type": "array"
          },
          "contact_saved": {
            "title": "Contact Saved",
            "type": "boolean"
          },
          "next_steps": {
            "items": {
              "enum": [
                "save_tracking_code",
                "check_status_later",
                "see_escalation_guidance"
              ],
              "type": "string"
            },
            "title": "Next Steps",
            "type": "array"
          },
          "published": {
            "const": false,
            "title": "Published",
            "type": "boolean"
          },
          "status": {
            "const": "received",
            "title": "Status",
            "type": "string"
          },
          "tracking_code": {
            "title": "Tracking Code",
            "type": "string"
          }
        },
        "required": [
          "tracking_code",
          "status",
          "published",
          "contact_saved",
          "attachments",
          "next_steps"
        ],
        "title": "ReportReceipt",
        "type": "object"
      },
...
      "StatusEventOut": {
        "properties": {
          "actor_type": {
            "title": "Actor Type",
            "type": "string"
          },
          "event_id": {
            "format": "uuid",
            "title": "Event Id",
            "type": "string"
          },
          "internal_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Internal Reason"
          },
          "new_status": {
            "title": "New Status",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "previous_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Status"
          },
          "public_message": {
            "title": "Public Message",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "previous_status",
          "new_status",
          "public_message",
          "actor_type",
          "occurred_at",
          "internal_reason"
        ],
        "title": "StatusEventOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ShaidaGo platform API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/reports": {
      "post": {
        "operationId": "reports_submit",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "attachments": {
                    "items": {
                      "format": "binary",
                      "type": "string"
                    },
                    "maxItems": 3,
                    "type": "array"
                  },
                  "concern_category": {
                    "enum": [
                      "no_visible_work",
                      "incomplete_work",
                      "unsafe_construction",
                      "suspected_incorrect_status",
                      "access_barrier",
                      "other_concern"
                    ],
                    "type": "string"
                  },
                  "contact_channel": {
                    "enum": [
                      "email",
                      "phone",
                      "messaging_app"
                    ],
                    "type": "string"
                  },
                  "contact_value": {
                    "maxLength": 200,
                    "type": "string"
                  },
                  "description": {
                    "maxLength": 8000,
                    "minLength": 10,
                    "type": "string"
                  },
                  "project_slug": {
                    "maxLength": 120,
                    "type": "string"
                  },
                  "reporter_handle": {
                    "maxLength": 200,
                    "type": "string"
                  },
                  "reporter_passphrase": {
                    "maxLength": 200,
                    "type": "string"
                  }
                },
                "required": [
                  "project_slug",
                  "concern_category",
                  "description"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Missing or malformed Idempotency-Key."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Reporter handle credentials not accepted."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Key reused with a different request."
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Content Too Large"
          },
          "415": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Submit Report",
        "tags": [
          "reports"
        ]
      }
    },
...
    "/v1/reviewer/reports": {
      "get": {
        "operationId": "reviewer_reports_queue",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "enum": [
                      "received",
                      "needs_information",
                      "under_review",
                      "verified_for_public_update",
                      "referred",
                      "closed"
                    ],
                    "type": "string"
                  },
                  "maxItems": 6,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "risk_level",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "standard",
                    "elevated",
                    "high"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk Level"
            }
          },
          {
            "in": "query",
            "name": "concern_category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "no_visible_work",
                    "incomplete_work",
                    "unsafe_construction",
                    "suspected_incorrect_status",
                    "access_barrier",
                    "other_concern"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Concern Category"
            }
          },
          {
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "minLength": 1,
                  "pattern": "^[a-z0-9-]+$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 1000,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueuePageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid pagination cursor."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Queue",
        "tags": [
          "reviewer"
        ]
      }
    },
    "/v1/reviewer/reports/{report_id}": {
      "get": {
        "operationId": "reviewer_reports_get",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Report Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_contact",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Contact",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid pagination cursor."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Detail",
        "tags": [
          "reviewer"
        ]
      }
    },
...
  }
}